requirements.txt
numpy==1.19.1
torch==1.9.0a0+2ecb2c7
torchvision==0.9.0a0
pillow==8.1.1
opencv-python==4.2.0.34
paddlepaddle==2.1.2
paddlehub==2.1.0
gradio==2.2.8
python-dotenv
python-decouple
python-multipart
fastapi 
uvicorn
